+++ include/minirt.h
@@ -6,7 +6,7 @@
 /*   By: iherman- <iherman-@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/08 15:28:02 by iherman-          #+#    #+#             */
-/*   Updated: 2025/08/25 19:08:30 by iherman-         ###   ########.fr       */
+/*   Updated: 2025/08/27 18:21:42 by iherman-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,11 +49,6 @@ typedef struct s_vec3
 	double	z;
 }	t_vec3;
 
-typedef struct s_ray
-{
-	t_vec3	origin;
-	t_vec3	direction;
-}	t_ray;
 
 typedef struct s_color
 {
@@ -61,6 +56,13 @@ typedef struct s_color
 	int		G;
 	int		B;
 }	t_color;
+
+typedef struct s_ray
+{
+	t_color	color;
+	t_vec3	origin;
+	t_vec3	direction;
+}	t_ray;
 
 typedef	struct s_amb_light
 {
@@ -72,6 +74,8 @@ typedef struct s_camera
 {
 	t_vec3	viewpoint;
 	t_vec3	norm_vec;
+	t_vec3	up;
+	t_vec3	right;
 	double	fov;
 }	t_camera;
 
@@ -106,6 +110,7 @@ typedef	struct s_cylinder
 
 typedef struct s_rt_data
 {
+	mlx_t	*mlx_win;
 	t_amb_light	amb_light;
 	t_camera	camera;
 	t_light		light;
@@ -128,5 +133,16 @@ void 	ft_control_type_identifier(char *file, t_rt_data *data);
 double	ft_prepare_to_convert_atoi(const char *line, int *i, t_rt_data *data);
 double	ft_prepare_to_convert_atof(const char *line, int *i, t_rt_data *data);
 // void	ft_control_line_end(char *line, int *i, t_rt_data *data);
+
+// vector math
+t_vec3	normalize(t_vec3 vec);
+double	vector_length_squared(t_vec3 vec);
+double	vector_length(t_vec3 vec);
+t_vec3	vector_add(t_vec3 vec1, t_vec3 vec2);
+t_vec3	vector_subtract(t_vec3 vec1, t_vec3 vec2);
+t_vec3	vector_multiply(t_vec3 vec, const double factor);
+bool	vector_is_larger(t_vec3 larger, t_vec3 smaller);
+t_vec3	vector_cross(t_vec3 vec1, t_vec3 vec2);
+double	vector_dot(t_vec3 vec1, t_vec3 vec2);
 
 #endif // MINIRT_H